# docs/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    RoleAssignmentOut:
      description: 'A grant of a role to a console account.


        `granted_at` is the assignment''s creation time; `updated_at` backs the

        `updated_since` reconciliation feed. Both the user and role are embedded so

        an IGA connector can reconcile without extra lookups.'
      properties:
        granted_at:
          description: When the role was granted (provisioned).
          format: date-time
          title: Granted At
          type: string
        id:
          title: Id
          type: integer
        role:
          $ref: '#/components/schemas/RoleAssignmentRoleRef'
        updated_at:
          format: date-time
          title: Updated At
          type: string
        user:
          $ref: '#/components/schemas/RoleAssignmentUserRef'
      required:
      - id
      - user
      - role
      - granted_at
      - updated_at
      title: RoleAssignmentOut
      type: object
    RoleAssignmentRoleRef:
      description: The role side of an assignment, embedded in assignment responses.
      properties:
        id:
          title: Id
          type: integer
        is_active:
          title: Is Active
          type: boolean
        name:
          title: Name
          type: string
      required:
      - id
      - name
      - is_active
      title: RoleAssignmentRoleRef
      type: object
    RoleAssignmentUserRef:
      description: The console-account side of an assignment, embedded in responses.
      properties:
        id:
          title: Id
          type: integer
        username:
          title: Username
          type: string
      required:
      - id
      - username
      title: RoleAssignmentUserRef
      type: object
...
    RoleGrant:
      description: Request body to grant (provision) a role to a console account.
      properties:
        role_id:
          description: ID of the role to grant.
          title: Role Id
          type: integer
      required:
      - role_id
      title: RoleGrant
      type: object
...
    UserCreate:
      description: Request body for creating a local console account.
      properties:
        password:
          maxLength: 255
          minLength: 8
          title: Password
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
          default: view_only
        username:
          maxLength: 100
          minLength: 1
          title: Username
          type: string
      required:
      - username
      - password
      title: UserCreate
      type: object
    UserOut:
      description: A console account as returned by the users API (never the password hash).
      properties:
        auth_type:
          description: '''local'' (password login) or ''sso'' (provisioned via an identity provider).'
          readOnly: true
          title: Auth Type
          type: string
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          title: Id
          type: integer
        is_active:
          title: Is Active
          type: boolean
        is_seeded:
          title: Is Seeded
          type: boolean
        last_login_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Login At
        role:
          title: Role
          type: string
        username:
          title: Username
          type: string
      required:
      - id
      - username
      - role
      - is_active
      - is_seeded
      - created_at
      - last_login_at
      - auth_type
      title: UserOut
      type: object
    UserRole:
      description: "Authorization role governing what a user can do in the web UI.\n\n- ADMIN: full access,\
        \ including settings and lookup management.\n- MANAGEMENT: full employee CRUD; can view (but not\
        \ manage) lookups and\n  the activity log; no settings access.\n- VIEW_ONLY: read-only access\
        \ to employees and the activity log.\n\nRoles govern the web UI only; the REST API is authorized\
        \ separately via\nAPI keys and OAuth clients."
      enum:
      - admin
      - management
      - view_only
      title: UserRole
      type: string
    UserUpdate:
      description: Request body for updating a console account. All fields optional.
      properties:
        password:
          anyOf:
          - maxLength: 255
            minLength: 8
            type: string
          - type: 'null'
          title: Password
        role:
          anyOf:
          - $ref: '#/components/schemas/UserRole'
          - type: 'null'
        username:
          anyOf:
          - maxLength: 100
            minLength: 1
            type: string
          - type: 'null'
          title: Username
      title: UserUpdate
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Lightweight HR Source of Truth for Saviynt Identity Cloud POC and integration testing.
  title: Demo HR Source of Truth App
  version: 1.7.2
openapi: 3.1.0
paths:
...
  /api/v1/users/:
    get:
      description: List all console accounts.
      operationId: list_users_api_v1_users__get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserOut'
                title: Response List Users Api V1 Users  Get
                type: array
          description: Successful Response
      summary: List Users
      tags:
      - users
    post:
      description: Create a local (password) console account.
      operationId: create_user_api_v1_users__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create User
      tags:
      - users
  /api/v1/users/{user_id}:
    get:
      description: Get a single console account by ID.
      operationId: get_user_api_v1_users__user_id__get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get User
      tags:
      - users
    patch:
      description: Update a console account's username, password, and/or role.
      operationId: update_user_api_v1_users__user_id__patch
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update User
      tags:
      - users
  /api/v1/users/{user_id}/disable:
    post:
      description: Disable a console account so it can no longer sign in.
      operationId: disable_user_api_v1_users__user_id__disable_post
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Disable User
      tags:
      - users
  /api/v1/users/{user_id}/enable:
    post:
      description: Re-enable a disabled console account.
      operationId: enable_user_api_v1_users__user_id__enable_post
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Enable User
      tags:
      - users
  /api/v1/users/{user_id}/roles:
    get:
      description: List the access roles currently granted to a console account.
      operationId: list_user_roles_api_v1_users__user_id__roles_get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RoleAssignmentOut'
                title: Response List User Roles Api V1 Users  User Id  Roles Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List User Roles
      tags:
      - users
    post:
      description: Grant (provision) an access role to a console account.
      operationId: grant_user_role_api_v1_users__user_id__roles_post
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleGrant'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignmentOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Grant User Role
      tags:
      - users
  /api/v1/users/{user_id}/roles/{role_id}:
    delete:
      description: Revoke (deprovision) an access role from a console account.
      operationId: revoke_user_role_api_v1_users__user_id__roles__role_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: integer
      - in: path
        name: role_id
        required: true
        schema:
          title: Role Id
          type: integer
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke User Role
      tags:
      - users
...